+++ modules/src/ics02_client/msgs/update_client.rs
@@ -44,11 +44,6 @@ impl Msg for MsgUpdateAnyClient {
         crate::keys::ROUTER_KEY.to_string()
     }
 
-    fn validate_basic(&self) -> Result<(), Self::ValidationError> {
-        // Nothing to validate since all fields are validated on creation.
-        Ok(())
-    }
-
     fn type_url(&self) -> String {
         TYPE_URL.to_string()
     }