+++ modules/src/ics04_channel/mod.rs
@@ -1,7 +1,11 @@
 //! ICS 04: IBC Channel implementation
 
 pub mod channel;
+pub mod context;
 pub mod error;
 pub mod events;
+
+pub mod handler;
 pub mod msgs;
 pub mod packet;
+pub mod version;