+++ modules/src/ics04_channel/msgs/timeout.rs
@@ -46,12 +46,6 @@ impl Msg for MsgTimeout {
         crate::keys::ROUTER_KEY.to_string()
     }
 
-    fn validate_basic(&self) -> Result<(), Self::ValidationError> {
-        // Nothing to validate
-        // All the validation is performed on creation
-        Ok(())
-    }
-
     fn type_url(&self) -> String {
         TYPE_URL.to_string()
     }