+++ modules/src/ics26_routing/handler.rs
@@ -8,10 +8,11 @@ use crate::ics02_client::msgs::create_client;
 use crate::ics02_client::msgs::update_client;
 use crate::ics02_client::msgs::ClientMsg;
 use crate::ics03_connection::handler::dispatch as ics3_msg_dispatcher;
+use crate::ics04_channel::handler::dispatch as ics4_msg_dispatcher;
 use crate::ics26_routing::context::ICS26Context;
 use crate::ics26_routing::error::{Error, Kind};
 use crate::ics26_routing::msgs::ICS26Envelope;
-use crate::ics26_routing::msgs::ICS26Envelope::{ICS2Msg, ICS3Msg};
+use crate::ics26_routing::msgs::ICS26Envelope::{ICS2Msg, ICS3Msg, ICS4Msg};
 
 /// Mimics the DeliverTx ABCI interface, but a slightly lower level. No need for authentication
 /// info or signature checks here.
@@ -94,7 +95,21 @@ where
                 .with_log(handler_output.log)
                 .with_events(handler_output.events)
                 .with_result(())
-        } // TODO: add dispatchers for ICS4 and others.
+        }
+
+        ICS4Msg(msg) => {
+            let handler_output =
+                ics4_msg_dispatcher(ctx, msg).map_err(|e| Kind::HandlerRaisedError.context(e))?;
+
+            // Apply any results to the host chain store.
+            ctx.store_channel_result(handler_output.result)
+                .map_err(|e| Kind::KeeperRaisedError.context(e))?;
+
+            HandlerOutput::builder()
+                .with_log(handler_output.log)
+                .with_events(handler_output.events)
+                .with_result(())
+        } // TODO: add dispatchers for others.
     };
 
     Ok(output)
@@ -114,6 +129,13 @@ mod tests {
     use crate::ics03_connection::msgs::conn_open_try::test_util::get_dummy_msg_conn_open_try;
     use crate::ics03_connection::msgs::conn_open_try::MsgConnectionOpenTry;
     use crate::ics03_connection::msgs::ConnectionMsg;
+
+    use crate::ics04_channel::msgs::chan_open_init::test_util::get_dummy_raw_msg_chan_open_init;
+    use crate::ics04_channel::msgs::chan_open_init::test_util::get_dummy_raw_msg_chan_open_init_with_missing_connection;
+
+    use crate::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
+    use crate::ics04_channel::msgs::ChannelMsg;
+
     use crate::ics24_host::identifier::ClientId;
     use crate::ics26_routing::handler::dispatch;
     use crate::ics26_routing::msgs::ICS26Envelope;
@@ -153,6 +175,14 @@ mod tests {
         let msg_conn_try_good_height =
             MsgConnectionOpenTry::try_from(get_dummy_msg_conn_open_try(10, 29)).unwrap();
 
+        let msg_chan_init =
+            MsgChannelOpenInit::try_from(get_dummy_raw_msg_chan_open_init()).unwrap();
+
+        let msg_chan_init2 = MsgChannelOpenInit::try_from(
+            get_dummy_raw_msg_chan_open_init_with_missing_connection(),
+        )
+        .unwrap();
+
         // We reuse this same context across all tests. Nothing in particular needs parametrizing.
         let mut ctx = MockContext::default();
 
@@ -169,6 +199,8 @@ mod tests {
             create_client_msg,
             res
         );
+
+        ctx.add_port(msg_chan_init.port_id().clone());
 
         // Figure out the ID of the client that was just created.
         // TODO: Create a "search by attribute key" API for HandlerOutput to simplify the following
@@ -236,6 +268,17 @@ mod tests {
                 msg: ICS26Envelope::ICS3Msg(ConnectionMsg::ConnectionOpenTry(Box::new(
                     msg_conn_try_good_height,
                 ))),
+                want_pass: false,
+            },
+            // ICS04
+            Test {
+                name: "Channel open init success".to_string(),
+                msg: ICS26Envelope::ICS4Msg(ChannelMsg::ChannelOpenInit(msg_chan_init)),
+                want_pass: true,
+            },
+            Test {
+                name: "Channel open init fail due to missing connection".to_string(),
+                msg: ICS26Envelope::ICS4Msg(ChannelMsg::ChannelOpenInit(msg_chan_init2)),
                 want_pass: false,
             },
         ]