+++ relayer/src/error.rs
@@ -18,10 +18,6 @@ pub enum Kind {
     #[error("I/O error")]
     Io,
 
-    /// Poisoned mutex
-    #[error("poisoned mutex")]
-    PoisonedMutex,
-
     /// Invalid configuration
     #[error("Invalid configuration")]
     Config,